README.md
This is an android application that demonstrates pulling json strings from an api for stock info. saving this data to a file and then reading it back in. uses fragments to allow for adaptive UI based on device state. (ie. landscape or portrait).
To use app must enter company stock symbols e.g. AAPL for Apple Inc. or FB for facebook.